:sparkles: marking day4 UI fixes and bug fixes
