+++ app/src/main/AndroidManifest.xml
@@ -9,6 +9,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
+        <activity android:name=".Marking3Activity"></activity>
         <activity
             android:name=".Main3Activity"
             android:label="@string/title_activity_main3"
