+++ app/src/main/res/values/colors.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <color name="colorPrimary">#6200EE</color>
-    <color name="colorPrimaryDark">#3700B3</color>
+    <color name="colorPrimary">#feb63d</color>
+    <color name="colorPrimaryDark">#feb63d</color>
     <color name="colorAccent">#03DAC5</color>
 
     <color name="black_overlay">#66000000</color>
