+++ app/src/main/res/values/strings.xml
@@ -18,7 +18,7 @@
 
     <string name="page_name3">Save your Data</string>
     <string name="subject_name">Subject name :-</string>
-    <string name="id">Id :-</string>
+    <string name="id">Subject Id :-</string>
     <string name="btn_add">ADD DATA</string>
     <string name="btn_update">UPDATE</string>
     <string name="btn_delete">DELETE</string>
